qa_test/pyproject.toml
[project]
name = "qa-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.190.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "playwright>=1.58.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
